Move logging logic to a new class.
