+++ HTTPServer.driver/BadRequestFilter.cs
@@ -1,5 +1,5 @@
 ﻿using System.Text;
-using HTTPServer.core;
+using HTTPServer.Core;
 using System;
 using System.IO;
 
@@ -25,28 +25,9 @@ namespace HTTPServer.app
                 return reply;
             }
             var returnedReply = _nextCommand.Execute(request);
-            LogData(LogMessage(request, returnedReply));
+            Logger logger = new Logger();
+            logger.LogData(logger.LogMessage(request, returnedReply));
             return returnedReply;
-        }
-        
-        private void LogData(string loggingMessage)
-        {
-            Console.Write(loggingMessage);
-        }
-
-        private string LogMessage(Request request, Reply reply)
-        {
-            var responseCode = GetResponseCode(reply.StartingLine);
-            var logMessage = DateTime.Now.ToString() + " " + request.Method + " " + request.Uri + " "
-                + request.HttpVersion + " " + responseCode + "\r\n";
-            return logMessage;
-        }
-
-        public string GetResponseCode(byte[] response)
-        {
-            var firstLine = Encoding.UTF8.GetString(response);
-            var requestLine = firstLine.Split(' ', ' ');
-            return requestLine[1];
         }
 
         private bool IsMalformed(Request request)
